requirements.txt
langgraph>=0.0.19
langchain>=0.1.0
langchain-core>=0.1.0
langchain-openai>=0.0.1
openai>=1.0.0
pydantic>=2.0.0
pyyaml>=6.0
python-dotenv>=1.0.0
requests>=2.31.0